No longer storing graph data in redux. Graph is constructed in the ClusterGraphView component.
